Объединяет списки фильмов в 1 компонент, теперь генерирует количество фильмов в main.js
