+++ src/components/films-list.js
@@ -1,23 +1,29 @@
 import {createElement} from '../utils.js';
 
 /**
- * создаёт и возвращает разметку для раздела с фильмами
+ * Создаёт и возвращает разметку для раздела с фильмами
+ * @param {Object} filmList - данные из объекта списков с фильмами
  * @return {String}
  */
-const getFilmsListTemplate = () => (/* html */
-  `<section class="films-list">
-    <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
-    <div class="films-list__container"></div>
-  </section>`
-);
+const getFilmsListTemplate = (filmList) => {
+  const {title, extra, sort} = filmList;
+
+  return (/* html */
+    `<section class="${extra ? `films-list--extra` : `films-list`}">
+      <h2 class="films-list__title ${extra ? `visually-hidden` : ``}">${title}</h2>
+      <div class="films-list__container"></div>
+    </section>`
+  );
+};
 
 export default class FilmList {
-  constructor() {
+  constructor(filmLists) {
+    this._filmLists = filmLists;
     this._element = null;
   }
 
   getTemplate() {
-    return getFilmsListTemplate();
+    return getFilmsListTemplate(this._filmLists);
   }
 
   getElement() {
