+++ src/utils.js
@@ -3,35 +3,6 @@ export const RenderPosition = {
   BEFOREEND: `beforeend`
 };
 
-/**
- * Возращает случайный индекс массива
- * @param {Array} array - данные из массива
- * @return {Number}
- */
-export const getRandomIndex = (array) => {
-  return Math.floor(Math.random() * array.length);
-};
-
-/**
- * Возращает случайное булево значение
- * @return {Boolean}
- */
-export const getRandomBoolean = () => {
-  return Math.random() >= 0.5;
-};
-
-/**
- * Возращает случайный элемент из массива
- * @param {Array} array - данные из массива
- * @return {String}
- */
-export const getRandomElementFromArray = (array) => {
-  const randomIdx = getRandomIndex(array);
-  const randomElement = array.find((title, idx) => idx === randomIdx);
-
-  return randomElement;
-};
-
 export const createElement = (template) => {
   const newElement = document.createElement(`div`);
   newElement.innerHTML = template;
